MVP
Intakes data in the form of json text files. Each muscle and exercise has a defined object. Added a filter function to get specific exercises based on equipment or form. Added a plot function to plot exercise progression graphs. Add a help function to help orientate when launching the .exe